minor code renaming and refactor, added datgui and some settings to change
