+++ src/domElements.ts
@@ -1,9 +1,8 @@
 import { FullWindowSize } from "./types";
 
+// initial canvas setup and global exports
 export const canvas = document.querySelector<HTMLCanvasElement>('canvas')!;
 export const ctx = canvas.getContext('2d')!;
-
-// initial canvas setup
 
 // initial fullscreen
 canvas.width = FullWindowSize.getWidth();
